### 2020-11-11 #2834 @whedon
[PRE REVIEW]: deeplenstronomy: A dataset simulation package for strong gravitational lensing
**Submitting author:** @rmorgan10 (<a href="http://orcid.org/0000-0002-7016-5471">Robert Morgan</a>)
**Repository:** <a href="https://github.com/deepskies/deeplenstronomy" target ="_blank">https://github.com/deepskies/deeplenstronomy</a>
**Version:** v0.0.1.0
