### 2020-11-11 #2834 @kthyng
@rmorgan10 not sure about the error in your paper... can you compare your line 9 with the [paper requirements](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) and see if something is amiss?
