### 2020-11-11 #2834 @rmorgan10
It looks like lines 8-9 are
```
authors:
  - name: Robert Morgan^[Corresponding author email: robert.morgan@wisc.edu]
```
Though I believe I followed the same syntax as the paper requirements:
```
authors:
  - name: Adrian M. Price-Whelan^[Custom footnotes for e.g. denoting who the corresponding author is can be included like this.]
```

I'm not quite sure what the problem is either.
