### 2020-11-11 dc8e4aa
try removing footnote
