### 2020-11-11 #2834 @rmorgan10
I have a feeling the compile issue is the ':' character in the footnote, or perhaps that the footnote text needs to be enclosed in single quotes. I will test this as soon as I can, but at present the online compile service seems to be down.
