### 2020-11-12 #2834 @arfon
> I have a feeling the compile issue is the ':' character in the footnote, or perhaps that the footnote text needs to be enclosed in single quotes. I will test this as soon as I can, but at present the online compile service seems to be down.

This should be fixed again now.

@rmorgan10 - is this JOSS submission associated with a paper in AAS publishing?
