### 2020-11-12 #2834 @pdebuyl
@rmorgan10 the AAS question should be answered before I proceed. BTW, as you have an ORCID record the email could be considered superfluous, which would solve the compilation issue :-)
