+++ paper.bib
@@ -64,7 +64,7 @@
       primaryClass={astro-ph.CO}
 }
 
-@article{astropy
+@article{astropy,
 Adsnote = {Provided by the SAO/NASA Astrophysics Data System},
 Adsurl = {http://adsabs.harvard.edu/abs/2013A%26A...558A..33A},
 Archiveprefix = {arXiv},
