### 2020-11-12 c25c06e
remove email from footnote
