### 2020-11-12 7250a9f
move figure lower to avoid big whitespace
