### 2020-11-12 #2834 @rmorgan10
@arfon this JOSS submission is not associated with a paper in a AAS journal.

@pdebuyl , @kthyng the compilation issue has been fixed and everything looks good!
