### 2020-11-13 #2834 @pdebuyl
@fipanther @shreyasbapat can you review the article "deeplenstronomy: A dataset simulation package for strong gravitational lensing" for JOSS ?
