### 2020-11-13 #2834 @rmorgan10
`deeplenstronomy` doesn't perform any gravitational lensing calculations itself, it's a wrapper that focuses on generating and organizing data for deep learning using the package [`lenstronomy`](https://github.com/sibirrer/lenstronomy) (which is the main gravitational lensing workhorse).

After a quick glance at the @einsteinpy family, to me it doesn't feel like `deeplenstronomy` fits in, but I'd be happy to discuss it more with you if you think it does!
