### 2020-11-13 a612f73
minor updates for new features
