### 2020-11-14 #2834 @pdebuyl
@kakirastern I pick reviewers from this list first and try to balance the reviewers' backgrounds. The best way forward, in any case, is to volunteer to review for JOSS here: https://joss.theoj.org/reviewer-signup.html Then you can be picked for a future submission.
