+++ paper.bib
@@ -41,7 +41,8 @@
   volume={445},
   pages={51--56},
   year={2010},
-  organization={Austin, TX}
+  organization={Austin, TX},
+  doi = {10.25080/majora-92bf1922-00a}
 }
 
 @article{matplotlib,
@@ -55,13 +56,17 @@
     doi = {10.1109/MCSE.2007.55}
 }
 
-@misc{lenstronomy,
-      title={Lenstronomy: multi-purpose gravitational lens modelling software package}, 
-      author={Simon Birrer and Adam Amara},
-      year={2018},
-      eprint={1803.09746},
-      archivePrefix={arXiv},
-      primaryClass={astro-ph.CO}
+@article{lenstronomy,
+title = "lenstronomy: Multi-purpose gravitational lens modelling software package",
+journal = "Physics of the Dark Universe",
+volume = "22",
+pages = "189 - 201",
+year = "2018",
+issn = "2212-6864",
+doi = "10.1016/j.dark.2018.11.002",
+url = "http://www.sciencedirect.com/science/article/pii/S2212686418301869",
+author = "Simon Birrer and Adam Amara",
+keywords = "Gravitational lensing, Software, Image simulations"
 }
 
 @article{astropy,
