### 2020-11-21 #2854 @rmorgan10
> There are a few DOIs missing !

The `pandas` DOI has been added, and the `lenstronomy` citation has been updated to the journal article (including the DOI) instead of ArXiv.
