### 2020-11-30 #2854 @pdebuyl
Thank you for the reply @jiwoncpark . Since the spring, the timeframe for a JOSS review is 6 weeks maximum. I update the review regularly to make sure that the review will happen but there is no emergency for tomorrow.
