### 2020-12-08 a97fd8c
fix bug in storing of theta_E and sigma_v
