### 2020-12-10 #2854 @rmorgan10
@pdebuyl I would like to put the `deeplenstronomy` paper on arXiv at some point, but this is the first time I've worked with JOSS. Would you recommend waiting until after the review process for any reason?
