### 2020-12-10 #2854 @pdebuyl
Hi @rmorgan10 I would recommend waiting for the publication here to save you the trouble of updating the arXiv version. You are free to proceed however (see https://joss.readthedocs.io/en/latest/submitting.html?highlight=arxiv#preprint-policy ).

Just make sure to not reference the paper as published in joss yet (mention of submission is fine of course).
