### 2020-12-10 #2854 @rmorgan10
@pdebuyl If I wait until after the review process (and `deeplenstronomy` is accepted), does JOSS release the TeX for the paper to me?
