### 2020-12-10 #2854 @pdebuyl
No. Officially, JOSS only sees the input .md file and the pdf output. It is possible to execute our code locally of course. It uses pandoc. It is also possible to use pandoc directly, provided you have the template and logo. I can help if you wish (not anymore today however).
