### 2020-12-10 #2854 @arfon
@rmorgan10 - you might be able to make a TeX version using this script: https://github.com/mattpitkin/psrqpy/blob/master/paper/Makefile
