### 2020-12-14 61963d0
Add citation guidelines to README
