### 2020-12-16 deepskies/deeplenstronomy#41 @jiwoncpark
Improving documentation (JOSS review)
These are some minor suggestions on improving the documentation in the README. openjournals/joss-reviews#2854

- (optional) It'd be great if the README could point directly to the docs page, where the user can read the per-module docstrings. The "Generating Datasets" has the parameter description for `make_dataset()` in the markdown, but I wouldn't have known to look for it there.
- "Getting Started" notebook: link for "Creating deeplenstronomy Configuration Files"  is broken.
- Currently, all the tests are in `test_all.py` and it's a bit hard to tell exactly what functionality of what module is being tested. A better organization would be to break it down into multiple files named after the modules being tested. Also, I strongly recommend setting up CI via GitHub Actions!
