### 2020-12-16 #2854 @jiwoncpark
Hi @rmorgan10, I created an [issue](https://github.com/deepskies/deeplenstronomy/issues/41#issue-768442817) at the `deeplenstronomy` repo to address some of the review checklist items, including tests and function docstrings. Separately, I'm adding my comments on the paper manuscript below. @pdebuyl Please let me know if the issue or paper comments should be placed elsewhere.

**Paper comments**

*Background*
- "They also provided critical evidence during tests of the theory of General Relativity." --> would be better in present tense (SL as a standing test of GR and modified theories of gravity)
- "SDSS, DES, and LSST" --> I'm not sure about the JOSS rules on survey acronyms, but I wonder if these should be written out.
- You can introduce the 2D-CNN acronym after "two-dimensional convolutional neural network" but see my comments below.
- "an integral parts" --> "an integral part"

*Statement of need*
- "SL detection" : My impression is that this software is more general and can be used for lens modeling, time series analysis, etc. not just detection.
- I wouldn't restrict the architecture to 2D CNNs either, to make room for, e.g. graph-based architectures.

*Summary*
- "data set" --> "dataset" (multiple places)
- The second paragraph contains a long list, some elements of which don't quite correspond to "[making the] datasets as realistic as possible." Maybe break it up into several logically coherent lists? You could present "inspect and visualize the simulation outputs," "simulate time series data" as additional tools and functionalities, respectively.
- Can you elaborate on the time series simulation portion? You mainly discuss images before this point. After looking at a demo notebook, I understand this to mean a time series of images?
- A code snippet and a short yaml file would be helpful, as a preview of the API.

*State of the field*
Currently, there is no section or paragraph describing the "state of the field" (addressing the review criterion: "Do the authors describe how this software compares to other commonly-used packages?"). I'm not sure if this part of the JOSS requirements, but if I can mention my own software package, `baobab` (https://github.com/jiwoncpark/baobab) is a very similar `lenstronomy` wrapper with a focus on lens modeling. It was used to generate the training set for two modeling papers, [Park et al 2020](https://arxiv.org/abs/2012.00042) and [Wagner-Carena et al 2020](https://arxiv.org/abs/2010.13787). `deeplenstronomy` seems more general.

*Overall style comments*
- Sentences tend to be quite long, with one or more descriptive clauses. I recommend splitting them into shorter sentences.
- Using active voice and fewer nouns would also help with the above.
Example: "Due to the inherent dependence of the performance of ML approaches on their training data,the 2D-CNN approach to SL detection is in tension with scientific reproducibility without a clear prescription for the simulation of the training data." --> Because supervised ML approaches depend heavily on the training data, the principle of reproducibility dictates that they be accompanied by a clear ... (or something like that)
