### 2020-12-16 #2854 @pdebuyl
@rmorgan10 there is some confusion related to the link error mentioned in https://github.com/deepskies/deeplenstronomy/issues/41 : you have a repository called deeplenstronomy with the notebooks. Is that repository supposed to remain online? If so, could you rename it?
