### 2020-12-16 #2854 @rmorgan10
> Edit/PS to previous comment: "you have a repository called deeplenstronomy in your github profile"

Yeah that was a staging area for the notebooks before we enabled github-pages for the deepskies/deeplenstronomy repo. I should really just delete the personal repo. Thanks for pointing that out!
