### 2020-12-16 68ffd33
update hyperlinks in GettingStarted
