### 2020-12-16 deepskies/deeplenstronomy#41 @rmorgan10
Links in the "Getting Started" Notebook are now fixed by commit [68ffd33](https://github.com/deepskies/deeplenstronomy/commit/68ffd33a9efd3bd5ccc194f64ae6bdd5da5b946f)
