### 2020-12-17 #2854 @rmorgan10
@jiwoncpark Thanks for your review! I think you've pointed out things that will help improve `deeplenstronomy` a lot. I've started working on some of the small things in the issue you opened in the `deeplenstronomy` repo.

I just wanted to give everyone a heads up that I won't be focusing too much on implementing major changes from the JOSS reviews until I have some more free time (which for me will start Jan 5). So @shreyasbapat no need to rush to get your review in if you are also busy.

Thanks again!
