### 2021-01-11 #2854 @pdebuyl
Thanks for the reply @rmorgan10 . Indeed, you can wait for the second review (JOSS allows for both ways of working).
