### 2021-01-14 deepskies/deeplenstronomy#41 @shreyasbapat
I won't create a new issue for the test suite since it is already mentioned here :D
