### 2021-01-14 deepskies/deeplenstronomy#41 @shreyasbapat
I would also like to point out that even though the package has excellent docstrings pretty much everywhere, there's no simple way to see what the API looks like. I feel there is a need to have the API Documentation for the same. Which can be easily achieved using sphinx.
