### 2021-01-14 deepskies/deeplenstronomy#42 @shreyasbapat
Coverage of the test suite (JOSS Review)
The test suite has two major issues:
1) There's no logical separation to check which module is being tested.
2) The test suite seems very small. I am curious if that is getting you enough coverage to check everything. However, this is not a necessary requirement for the JOSS Review, but it would be great to understand how much of the code is tested.

https://github.com/openjournals/joss-reviews/issues/2854
