### 2021-01-14 #2854 @shreyasbapat
I think @jiwoncpark mentioned a majority of issues already. I created an issue which might not be very imporant from JOSS persective, but from a user's and maintainability perspective.

I feel once the tests and documentation is done, we would be good to go.

https://github.com/deepskies/deeplenstronomy/issues/42
