### 2021-01-15 #2854 @rmorgan10
Hi All, thanks for the reviews! If I understand correctly, it looks like the major required improvements are:

1. Making the doc strings available via sphinx
2. Organize the tests and clarify what is being tested
3. Minor edits to the paper text

I am happy to do all three and will hopefully have all issues resolved at the end of next week if not sooner. I am glad the code and its functionalities were satisfactory. Thanks again for your careful reviews of deeplenstronomy!
