### 2021-01-18 d9d97ff
update docstrings and generate html docs
