+++ deeplenstronomy/__init__.py
@@ -1,3 +1,5 @@
+"""Welcome to deeplenstronomy! Check out example notebooks here: https://deepskies.github.io/deeplenstronomy"""
+
 __author__ = 'Robert Morgan'
 __email__ = 'robert.morgan@wisc.edu'
 __version__ = '0.0.1'
