+++ deeplenstronomy/deeplenstronomy.py
@@ -1,4 +1,4 @@
-# Outer shell to do everything for you
+"""The main module for dataset generation."""
 
 import os
 import random
@@ -19,10 +19,11 @@ class Dataset():
         """
         Create a dataset. If config file or config dict is supplied, generate it.
 
-        :param config: yaml file specifying dataset characteristics 
-                       OR                                                                                                                                                                                                                                                                                                                                                                                                                                       pre-parsed y                       yaml file as a dictionary 
-        :param store: If true, the generated data is stored as attributes of this object
-        :param save: If true, the generated data is written to disk
+        Args:
+            config (str or dict, optional, default=None): name of yaml configuration file
+                specifying dataset characteristics or a pre-parsed yaml file as a dictionary 
+            store (bool, optional, default=True): store the generated data as attributes of this object
+            save (bool, optional, default=False): save the generated data to disk
         """
         
         if config:
@@ -31,10 +32,11 @@ class Dataset():
 
     def update_param(self, new_param_dict, configuration):
         """
-        Update single parameters to new values
+        Update single parameters to new values.
 
-        :param new_param_dict: {'param_1_name': new_value_1, 'param_2_name': new_value_2, ...}
-        :param configuration: like 'CONFIGURATION_1', 'CONFIGURATION_2', etc...
+        Args:
+            new_param_dict (dict): {'param_1_name': new_value_1, 'param_2_name': new_value_2, ...}
+            configuration (str): like 'CONFIGURATION_1', 'CONFIGURATION_2', etc...     
         """
         # Put in the updated values
         for new_param, new_value in new_param_dict.items():
@@ -44,16 +46,20 @@ class Dataset():
 
     def update_param_dist(self, new_param_dist_dict, configuration):
         """
-        Update the distribution from which a parameter is drawn
+        Update the distribution from which a parameter is drawn.
 
-        :param new_param_dist_dict: should look like this:
-            {'param_1_name': {'name': 'uniform', 
-                              'parameters': {'minimum': new_value_1, 
-                                             'maximum': new_value_2}},
-             'param_2_name': {'name': 'uniform', 
-                              'parameters': {'minimum': new_value_3, 
-                                             'maximum': new_value_4}}, ...}
-        :param configuration: like 'CONFIGURATION_1', 'CONFIGURATION_2', etc...   
+        Args:
+            new_param_dist_dict (dict): 
+              Should look like this
+              
+                  {'param_1_name': {'name': 'uniform',
+                                  'parameters': {'minimum': new_value_1,
+                                                 'maximum': new_value_2}},
+                  'param_2_name': {'name': 'uniform',
+                                  'parameters': {'minimum': new_value_3,
+                                                 'maximum': new_value_4}}, ...}   
+            
+            configuration (str):  like 'CONFIGURATION_1', 'CONFIGURATION_2', etc...
         """
         # Put in the updated distributions
         for new_param, new_dist_info in new_param_dist_dict.items():
@@ -127,7 +133,8 @@ class Dataset():
         """
         Using the dictionary stored in self.config_dict, make a new dataset
         
-        :param make_dataset_args: dict, arguments supplied to make_dataset when original dataset was generated
+        Args:
+            make_dataset_args (dict): arguments supplied to make_dataset when original dataset was generated
         """
         params = dict(**make_dataset_args)
         params['config'] = self.config_dict
@@ -138,12 +145,13 @@ class Dataset():
[…]
