### 2021-01-18 a960561
remove unneeded level in docs dir
