### 2021-01-18 deepskies/deeplenstronomy#41 @rmorgan10
I've created a docs page to make the docstrings more accessible with commit a960561096aa0e812deece8b3b34d30f883947de

https://deepskies.github.io/deeplenstronomy/docs
