### 2021-01-18 deepskies/deeplenstronomy#41 @rmorgan10
I've linked the docs page in the main readme so that users can easily find it with commit 7d1d0b71e94b939570b82fc1afbd3fa9def05aef
