### 2021-01-18 deepskies/deeplenstronomy#41 @rmorgan10
@jiwoncpark I'm planning on addressing your comments about the tests in @shreyasbapat 's issue #42 . When you get a chance, could you look over the improvements I've made to the documentation and let me know if you feel like this issue can be closed?
