### 2021-01-19 0593e7e
minor changes from JOSS review
