+++ paper.md
@@ -41,38 +41,40 @@ bibliography: paper.bib
 # Background
 
 Astronomical observations and statistical modeling permit the high-fidelity analysis of strong gravitational lensing (SL) systems, which display an astronomical phenomenon in which light from a distant object is deflected by the gravitational field of another object along its path to the observer.
-These systems are of great scientific interest, because they provide information about multiple astrophysical and cosmological phenomena, including the nature of dark matter, the expansion rate of the Universe, and characteristics of galaxy populations. 
-They also provided critical evidence during tests of the theory of General Relativity. 
+These systems are of great scientific interest because they provide information about multiple astrophysical and cosmological phenomena, including the nature of dark matter, the expansion rate of the Universe, and characteristics of galaxy populations. 
+They also serve as standing tests of the theory of General Relativity and modified theories of gravity. 
 
 Traditional searches for SL systems have involved time- and effort-intensive visual or manual inspection of images by humans to identify characteristic features --- like arcs, particular color combinations, and  object orientations. 
-However, a comprehensive search using the traditional approach is prohibitively expensive for large numbers of images, like those in cosmological surveys --- e.g., SDSS [@sdss], DES [@des], and LSST [@lsst]. 
+However, a comprehensive search using the traditional approach is prohibitively expensive for large numbers of images, like those in cosmological surveys --- e.g., the Sloan Digital Sky Survey [@sdss], the Dark Energy Survey [@des], and the Legacy Survey of Space and Time [@lsst]. 
 To automate the SL detection process, techniques based on machine learning (ML) are beginning to overtake traditional approaches for scanning  astronomical images. 
-In particular, deep learning techniques --- like two-dimensional convolutional neural networks --- have been the focus, but they require large sets of labeled images to train these models. 
+In particular, deep learning techniques have been the focus, but they require large sets of labeled images to train these models. 
 Because of the relatively low number of observed SL systems, simulated datasets of images are often needed. 
-Thus, the composition and production of these simulated datasets have become an integral parts of the SL detection process.
+Thus, the composition and production of these simulated datasets have become integral parts of the SL detection process.
 
 One of the premier tools for simulating and analyzing SL systems, `lenstronomy` [@lenstronomy], works by the user specifying the properties of the physical systems, as well as how they are observed (e.g., telescope and camera) through a `python`-based application programming interface (API) to generate a single image. 
 Generating populations of SL systems that are fit for neural network training requires additional infrastructure. 
 
 # Statement of need 
 
-Due to the inherent dependence of the performance of ML approaches on their training data, the 2D-CNN approach to SL detection is in tension with scientific reproducibility without a clear prescription for the simulation of the training data. 
+Due to the inherent dependence of the performance of ML approaches on their training data, the deep learning approach to SL detection is in tension with scientific reproducibility without a clear prescription for the simulation of the training data. 
 There is a critical need for a tool that simulates full datasets in an efficient and reproducible manner, while enabling the use of all the features of the `lenstronomy` simulation API. 
 Additionally, this tool should  simplify user interaction with `lenstronomy` and organize the simulations and associated metadata into convenient data structures for deep learning problems.
 
 
 # Summary
 
-`deeplenstronomy` generates SL data sets by organizing and expediting user interaction with `lenstronomy`. 
+`deeplenstronomy` generates SL datasets by organizing and expediting user interaction with `lenstronomy`. 
[…]
