### 2021-01-19 0533893
incorporate state of the field text
