### 2021-01-19 d03bc5e
Add state of the field references
