### 2021-01-19 deepskies/deeplenstronomy#41 @rmorgan10
@jiwoncpark I've also updated the paper text based on your recommendations with the following commits:

- d03bc5e881374ce3cdb964327bb3c210f8792580
- 05338933bbc651167d0d55dd4cd124d9fb228ad3
- 0593e7e766d2c29786682b194505d69d56aaf1fa
