### 2021-01-20 #2854 @pdebuyl
>    - A code snippet and a short yaml file would be helpful, as a preview of the API.

Including such content in the paper goes against the guidelines, we consider it as documentation and recommend its inclusion in the proper documentation of the package.
