### 2021-01-20 deepskies/deeplenstronomy#41 @pdebuyl
The modules distributions and surveys don't have docstrings, it would be good to add them.
