### 2021-01-20 #2854 @rmorgan10
Thanks @pdebuyl . Code snippets and yaml file excerpts are available in the provided example notebooks, which are referenced in the paper, so hopefully this comment is already addressed within the JOSS guidelines.

https://deepskies.github.io/deeplenstronomy/Notebooks/
