### 2021-01-20 deepskies/deeplenstronomy#41 @rmorgan10
> The modules distributions and surveys don't have docstrings, it would be good to add them.

@pdebuyl The reason I left docstrings out for those modules is they are internal functions, never called by the user. The only reason I included them in the documentation and didn't make them private methods was to provide a quickly accessible list of all possible surveys and distributions.

Given that information, do you think the docstrings are still necessary for those modules?
