### 2021-01-20 3e9e4dc
convert test script into an outer shell for specific tests
