### 2021-01-20 6b1ad6d
organize testing suite into separate modules
