### 2021-01-20 deepskies/deeplenstronomy#42 @rmorgan10
@shreyasbapat I've now introduced a logical organization in the test suite with commits 3e9e4dc17a04da02e7edbe766155e7e89fb4b097 and 6b1ad6dbb433c9f296d1071500b25be35c811d48

When each separate testing module is run, a detailed description of what is being tested is now printed to the screen. You can see the output of the test suite in the updated Run_Tests notebook: https://github.com/deepskies/deeplenstronomy/blob/master/test/Run_Tests.ipynb
