### 2021-01-20 deepskies/deeplenstronomy#42 @rmorgan10
Overall, coverage is difficult to assess for this particular software. A lot of the internal functions are drawing from distributions using `random`, `numpy`, and `scipy` or simulating images with `lenstronomy`, so those parts of my software that have been tested in those particular packages.

What the tests I include in this suite cover is mapping the values in the input configuration file to the outputs of the simulation. Thus, by verifying that the generated simulations match expectations in terms of dataset size, attributes, objects and configurations included, the inclusion of time-series information, and other parameters, the testing suite here accomplishes that goal.

Please let me know if you have any feedback! Thanks!
