### 2021-01-25 deepskies/deeplenstronomy#41 @pdebuyl
> Given that information, do you think the docstrings are still necessary for those modules?

Yes. But a minimal docstring is fine. Else, the only starting point to figure what a function does is its name or the source code, which is as good as not providing their listing.
