### 2021-01-25 e906f49
add docstrings to internal functions
