+++ deeplenstronomy/distributions.py
@@ -1,4 +1,14 @@
-"""Contains all available distributions."""
+"""Contains all available distributions. Utilize the functions in this module
+by using the `DISTRIBUTION` keyword in your configuration file. As an example:
+
+seeing:
+    DISTRIBUTION:
+        NAME: uniform # function name to call
+        PARAMETERS:
+            minimum: 0 # value to set for function argument
+            maximim: 6 # value to set for function argument
+
+"""
 
 import numpy as np
 import random
@@ -7,31 +17,109 @@ from scipy.stats import poisson
 ## Single parameter sampling distributions
 
 def uniform(minimum, maximum, bands=''):
+    """
+    Return a samle from a uniform probability distribution
+    on the interval [`minimum`, `maximum`]
+
+    Args:
+        minimum (float or int): The minimum of the interval to sample
+        maximum (float or int): The maximum of the interval to sample
+
+    Returns:
+        A sample of the specified uniform distribution for each band in the simulation
+    """
     draw = random.uniform(minimum, maximum)
     return [draw] * len(bands.split(','))
 
 def uniform_int(minimum, maximum, bands=''):
+    """
+    Return a samle from a uniform probability distribution
+    on the interval [`minimum`, `maximum`] rounded to the nearest integer
+
+    Args:
+        minimum (int): The minimum of the interval to sample
+        maximum (int): The maximum of the interval to sample
+
+    Returns:
+        A rounded sample of the specified uniform distribution for each band in the simulation
+    """
     draw = round(random.uniform(minimum, maximum))
     return [draw] * len(bands.split(','))
 
 def normal(mean, std, bands=''):
+    """
+    Return a samle from a normal probability distribution
+    with specifeid mean and standard deviation
+
+    Args:
+        mean (float or int): The mean of the normal distribution to sample
+        std (float or int): The standard deviation of the normal distribution to sample
+
+    Returns:
+        A sample of the specified normal distribution for each band in the simulation
+    """
     draw = np.random.normal(loc=mean, scale=std)
     return [draw] * len(bands.split(','))
 
 def lognormal(mean, sigma, bands=''):
+    """
+    Return a samle from a lognormal probability distribution
+    with specifeid mean and standard deviation
+
+    Args:
+        mean (float or int): The mean of the lognormal distribution to sample
+        sigma (float or int): The standard deviation of the lognormal distribution to sample 
+
+    Returns:
+        A sample of the specified lognormal distribution for each band in the simulation
+    """
     draw = np.random.lognormal(mean=mean, sigma=sigma)
     return [draw] * len(bands.split(','))
 
 def delta_function(value, bands=''):
+    """
+    Use a delta function to set a specific value. Alternatively you can directly set the
+    value of a parameter if it is going to be constant. This functionality is useful if
+    you want to avoid deleting the `DISTRIBUTION` entry for a parameter in your
+    configuration file.
+
+    Args:
+        value : The value to set for this parameter
+
+    Returns:
[…]
