### 2021-01-25 cefeb62
format docstring example
