### 2021-01-25 deepskies/deeplenstronomy#41 @rmorgan10
@pdebuyl I've added docstrings to the distributions.py, surveys.py, and check.py modules with commit e906f49eedec9f36b2ae2437efbbbd8cc78966c1 and I've regenerated all the docs pages.
