### 2021-01-27 #2854 @pdebuyl
Thank you for the notification @rmorgan10

@shreyasbapat @jiwoncpark can you go over the changes and update the opened issues at deeplenstronomy's repo and the checkboxes here?
