### 2021-01-28 #2854 @pdebuyl
Hi @shreyasbapat thanks for finishing the review!

Can the issue https://github.com/deepskies/deeplenstronomy/issues/42 be closed?
