### 2021-01-28 #2854 @jiwoncpark
The text and code look really good to me, @rmorgan10! I've checked off the boxes. Super minor suggestions/comments:

- In line 43, I believe the package name ```SLSprinkler``` doesn't have a space and how about "imputes" --> "adds?" I'm also confirming the correct citation for the package, but this shouldn't hold you back in the JOSS review.
- Just FYI, you can choose what files to ignore in test coverage. Useful for excluding config files, files consisting of external modules tested elsewhere
- You might want to consider setting up GH actions for CI.

Once again, congrats on  a very nice package!
