### 2021-01-28 5bf82ae
minor changes to SLSprinkler phrasing
