### 2021-01-28 #2854 @rmorgan10
Thanks @jiwoncpark! I've implemented your line 43 changes. Please let me know if there is a better citation for `SLSprinkler` or `baobab`. I am considering setting up GitHub actions for CI at some point, but that will be slightly down the road.
