### 2021-01-29 #2854 @pdebuyl
Thank you @jiwoncpark for the review!

@rmorgan10 I'll now proceed to some editorial checks before handing out the paper to the editor-in-chief (see https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance ).
