### 2021-01-29 #2854 @pdebuyl
Hi @rmorgan10 don't worry about the DOI. I do a manual check of the DOIs, the automatic feature of whedon is a helpful starting point but not the source of truth :-)
