### 2021-01-29 4ce3e30
prevent float precision errors in checking number of exposures with time series
