### 2021-01-29 6fa3adf
update change log for JOSS release
