### 2021-01-29 #2854 @jiwoncpark
Hi, please use this citation for `baobab`:
```
Ji Won Park. (2021, January 28). jiwoncpark/baobab: v0.1.2 (Version v0.1.2). Zenodo. http://doi.org/10.5281/zenodo.4476822
```
