+++ paper.bib
@@ -27,12 +27,15 @@
   urldate = {2021-1-19}
 }
 
-@online{baobab,
-  author = {Ji Won Park},
-  title = {baobab},
-  year = 2020,
-  url = {https://github.com/jiwoncpark/baobab},
-  urldate = {2021-1-19}
+@software{baobab,
+  author       = {Ji Won Park},
+  title        = {jiwoncpark/baobab: v0.1.2},
+  month        = jan,
+  year         = 2021,
+  publisher    = {Zenodo},
+  version      = {v0.1.2},
+  doi          = {10.5281/zenodo.4476822},
+  url          = {https://doi.org/10.5281/zenodo.4476822}
 }
 
 @online{deeplenstronomy,
