### 2021-01-29 #2854 @rmorgan10
@pdebuyl Here is the version number and archive DOI for the updated release of deeplenstronomy:

Version: v0.0.1.2
DOI: 10.5281/zenodo.4479712
