### 2021-01-30 #2854 @jiwoncpark
If it's not too late, here is the Zenodo for `SLSprinkler`: http://doi.org/10.5281/zenodo.4480392
