+++ paper.bib
@@ -19,12 +19,19 @@
   doi     = {10.1038/s41586-020-2649-2}
 }
 
-@online{lsstdescsl,
-  author = {Bryce Kalmbach and Ji Won Park},
-  title = {LSST DESC SL Sprinkler},
-  year = 2020,
-  url = {https://github.com/LSSTDESC/SLSprinkler},
-  urldate = {2021-1-19}
+@software{lsstdescsl,
+  author       = {Kalmbach, J. Bryce and
+                  Park, Ji Won and
+                  Wiesner, Matthew and
+                  Chiang, James},
+  title        = {{LSSTDESC/SLSprinkler: LSST DESC Strong Lensing 
+                   Sprinkler for Simulations}},
+  month        = jul,
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {v1.1.0},
+  doi          = {10.5281/zenodo.4480392},
+  url          = {https://doi.org/10.5281/zenodo.4480392}
 }
 
 @software{baobab,
