### 2021-02-02 e9ab97f
implement static timeseries model
