### 2021-02-02 deepskies/deeplenstronomy#42 @rmorgan10
@shreyasbapat Can I go ahead and close this issue?
