### 2021-02-02 55ef572
add peak parameter for timeseries
