### 2021-02-03 #2854 @pdebuyl
Hi @rmorgan10

In the paper, can you use explicity journal names instead of the bibtex macro. `\prd` -> `Physical Review D`, etc? This is for Abbott,  Astropy, Ivezić, York.

SciPy has a paper out, please update the citation. see https://www.scipy.org/citing.html
