### 2021-02-03 032f6a5
Use journal names instead of LaTeX macros and update scipy
