+++ paper.bib
@@ -53,11 +53,28 @@
   urldate = {2020-10-06}
 }
 
-@Misc{scipy,
-  author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
-  title =     {{SciPy}: Open source scientific tools for {Python}},
-  year =      {2001},
-  url = "http://www.scipy.org/"
+@article{scipy,
+  author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and
+            Haberland, Matt and Reddy, Tyler and Cournapeau, David and
+            Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and
+            Bright, Jonathan and {van der Walt}, St{\'e}fan J. and
+            Brett, Matthew and Wilson, Joshua and Millman, K. Jarrod and
+            Mayorov, Nikolay and Nelson, Andrew R. J. and Jones, Eric and
+            Kern, Robert and Larson, Eric and Carey, C J and
+            Polat, {\.I}lhan and Feng, Yu and Moore, Eric W. and
+            {VanderPlas}, Jake and Laxalde, Denis and Perktold, Josef and
+            Cimrman, Robert and Henriksen, Ian and Quintero, E. A. and
+            Harris, Charles R. and Archibald, Anne M. and
+            Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and
+            {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
+  title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific
+            Computing in Python}},
+  journal = {Nature Methods},
+  year    = {2020},
+  volume  = {17},
+  pages   = {261--272},
+  adsurl  = {https://rdcu.be/b08Wh},
+  doi     = {10.1038/s41592-019-0686-2}
 }
 
 @inproceedings{pandas,
@@ -103,7 +120,7 @@ Author = {{Astropy Collaboration} and {Robitaille}, T.~P. and {Tollerud}, E.~J. 
 Doi = {10.1051/0004-6361/201322068},
 Eid = {A33},
 Eprint = {1307.6212},
-Journal = {\aap},
+Journal = {Astronomy \& Astrophysics},
 Keywords = {methods: data analysis, methods: miscellaneous, virtual observatory tools},
 Month = oct,
 Pages = {A33},
@@ -188,7 +205,7 @@ Bdsk-Url-1 = {https://dx.doi.org/10.1051/0004-6361/201322068}}
          {Weinberg}, David H. and {Yanny}, Brian and {Yasuda}, Naoki and
          {SDSS Collaboration}},
         title = "{The Sloan Digital Sky Survey: Technical Summary}",
-      journal = {\aj},
+      journal = {The Astronomical Journal},
      keywords = {Cosmology: Observations, Instrumentation: Miscellaneous, Astrophysics},
          year = 2000,
         month = sep,
@@ -262,7 +279,7 @@ archivePrefix = {arXiv},
          {Zenteno}, A. and {Zhang}, B. and {Zhang}, Y. and {Zuntz}, J. and
          {Dark Energy Survey Collaboration}},
         title = "{Dark Energy Survey year 1 results: Cosmological constraints from galaxy clustering and weak lensing}",
-      journal = {\prd},
+      journal = {Physical Review D},
      keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2018,
         month = aug,
@@ -396,7 +413,7 @@ archivePrefix = {arXiv},
          {Wittman}, David and {Wolff}, Sidney C. and {Wood-Vasey}, W. Michael and
          {Wu}, Xiuqin and {Xin}, Bo and {Yoachim}, Peter and {Zhan}, Hu},
[…]
