### 2021-02-03 #2854 @rmorgan10
@pdebuyl I've updated the bibliography with your recommendations, thanks! https://github.com/deepskies/deeplenstronomy/commit/032f6a5aa7aee1df7d4b6e6d1851304dde8c05ab
