### 2021-02-03 #2854 @pdebuyl
@rmorgan10 please check the final pdf https://raw.githubusercontent.com/openjournals/joss-papers/7b45a09803a630144e588ffe8f36aa033e3dbd66/joss.02854/10.21105.joss.02854.pdf

The editor-in-chief in rotation will pick up the process from here. Thanks for submitting to JOSS :-)
